Log now updates when objects are detected
The sensitivity may need be tweaked. By the nature of the camera thread and Yolo model, objects can be detected multiple times within a second, so this makes printing objects detected difficult. I added a delay so it only prints every 10.0 seconds, but this can maybe be configured in the options menu.