Added functionality to overlay and model toggles
The model can now be turned off from running and turned back on. There is currently a "bug" where if you turn off the object detection before starting camera feed, and then start up the object detection, the program will crash because it never initiated the model. I don't know the best way to fix this yet, but I'm thinking of solutions.